Added tabs to our Currency Converter Project. / Added .mdf and .ldf to my .gitignore file.
One tab is the Currency Converter App. Another Tab is the currency master (Used to create currencies),
which holds a database table for all the currency values used in our currency combo boxes

Removed the Window border from the UI
Created UI for the Currency Master Tab.

Added a new folder "Database".

-- Added a Service-Based database file "CurrencyConverter.mdf" inside the "Database" folder.
(This file is not added in the repository.)

After some research, I discovered that .mdf / ldf files are not supposed to be added in git repositories. (Rookie mistake. Learning every day)
More info here: https://stackoverflow.com/questions/29008222/where-do-i-put-mdf-and-ldf-files-to-share-an-sql-script-through-git

-- Added .mdf and .ldf to my .gitignore file.

A file for the queries used to create the tables will be created later on.

Will continue to setup my back-end.