Added new SQL connection objects (SqlConnection, SqlCommand, SqlAdapter) Created 3 variables to store data from the database.
Added System.Configuration to references in order to use the Configuration Manager class to access our connection string.

Created new methods, BindStaticData and BindDatabaseData.
BindStaticData is the same code from the BindCurrency method.
BindDatabaseData is a method that uses the currency data in the database to populate our combo boxes.
It also creates a new row for the default comboBox value, similar to how the BindStaticData method does it.
However, the datatable Object is now being populated by the database (In the BindDatabaseData method) instead of hardcoded values.

Will now work on the button click methods.