Added Newtonsoft.JSON nuget package to deserialize Http response.
Removed unused Static Resource from App.xaml

In the convert click method, modified the code in the else block to assign the parsed values in
the FromAmount and ToAmount variables. Then used these variables to calculate the converted value.

Created new async method GetDataFromAPI which takes in a url as a string and attempts to asynchronously
execute an http request, serialize the response and return the object.

Created console program that prints a string of all the properties for currencies.
Output of this program is copied in order to use it for the Rate class in the Currency Converter WPF app.
Also used console program to print all lines of code that add the currencies and rates to the datatable used for the combo boxes.

Added methods cmbFromCurrency_SelectionChanged and cmbToCurrency_SelectionChanged to automatically calculate currencies when provided.
The convert Button is practically obsolete. (Will be replacing Convert Button with ComboBox to select the data source. Static/API/Database)

Currency converter now works by pulling data from API.

Future Plans:

Will be creating Queries to insert the values received from the http request into my Currency_Master table.
Will add currency full names in the future