Currency Converter APP is Complete. View changes.
Commented out the code inside the cmbToCurrency_SelectionChanged and cmbFromCurrency_SelectionChanged methods.
Need to sort out bugs but will go take a Unity Course.

Added a label and a combo box to select which data source to use for the currencies.
Options are Static data, Database and API.

Modified currency converter to validate the combo box value and execute the correct methods in order
to set the data source for the currency.

Fixed bug in database logic that assigned id instead of amount to the currency rates combo boxes values.

Added try catch for when opening a database connection.
Will update readme soon with info on how to setup your database.

(Thinking about adding a developer mode checkbox that eliminates the database option unless you know what you're doing.)
There's a lot of room for improvement for this project but for now, I will move on to another project.